pyproject.toml
[project]
name = "quant-trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.42.0",
    "arch>=7.2.0",
    "colorama>=0.4.6",
    "pandas-stubs>=2.3.0.250703",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "statsmodels>=0.14.5",
    "types-pyyaml>=6.0.12.20250809",
    "yfinance>=0.2.65",
    "vaderSentiment>=3.3.2",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.5.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "filterpy>=1.4.5",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = 1
